refactor: @type 절대경로 설정
