--- src/types/type.ts
@@ -1,10 +1,0 @@
-export interface ChattingRoom {
-  id: string;
-  chat: {
-    chatId: number;
-    to: string;
-    from: string;
-    content: string;
-    time: string;
-  }[];
-}
